Update the package version being checked for the apache config file rename.
The version referenced had been 1.3.46 but that does not appear to be far
enough back, for upgrades from v1.3.18-2 (current wheezy) or v1.3.25
(current precise) or even v1.3.40 (current jessie). This should effectively
do the rename of the ledgersmb.conf apache config file for any package
version prior to 1.3.47-1 as necessary. Version changed to 1.3.18-2.