Copy UI/js-src/lsmb from upstream v1.5.13 to debian/missing-sources/UI/js/lsmb.
This is to resolve at least some of the 'source-is-missing' Lintian errors that
have been coming up: https://lintian.debian.org/tags/source-is-missing.html.
At least some alledged missing source files are in UI/js-src/lsmb, so copy them
to the debian/missing-sources direcotory so that Lintian can see them.